add Slices
