+++ src/components/private-route/private-route.tsx
@@ -2,6 +2,7 @@ import { Navigate } from 'react-router-dom';
 import { AppRoute, AuthorizationStatus } from '../../const';
 import { useAppSelector } from '../../hooks';
 import { getToken } from '../../api/token';
+import { getIsAuth } from '../../store/selectors/user-selector';
 
 
 type PrivateRootProps = {
@@ -9,7 +10,7 @@ type PrivateRootProps = {
 };
 
 function PrivateRoot({children}: PrivateRootProps): JSX.Element {
-  const authStatus = useAppSelector((state) => state.user.authorizationStatus);
+  const authStatus = useAppSelector(getIsAuth);
 
   return authStatus === AuthorizationStatus.Auth || getToken() !== ''
     ? (
