+++ src/hooks/use-film-details.ts
@@ -5,12 +5,13 @@ import { fetchFullFilmAction } from '../store/api-actions/api-film-actions.ts';
 import { ErrorType } from '../types.ts';
 import { NOT_FOUND_MESSAGE } from '../api/const.ts';
 import { AppRoute } from '../const.ts';
+import { getCurrentFilm } from '../store/selectors/film-selector.ts';
 
 
 export function useFilmDetails() {
   const { id = '' } = useParams();
   const dispatch = useAppDispatch();
-  const { film } = useAppSelector((state) => state.fullFilm);
+  const film = useAppSelector(getCurrentFilm);
   const navigate = useNavigate();
 
   useEffect(() => {
