--- src/mocs/films-list.ts
@@ -1,86 +1,0 @@
-export const FilmsList = [
-  {
-    id: 1,
-    filmName: 'The Disaster Artist',
-    genre: 'Drama',
-    filmReleaseDate: 2017,
-    filmImg: 'https://static01.nyt.com/images/2017/12/01/arts/01disasterartist-web/01disasterartist-web-superJumbo.jpg',
-  },
-  {
-    id: 2,
-    filmName: 'Hit the Road',
-    genre: 'Comedy',
-    filmReleaseDate: 2021,
-    filmImg: 'https://assets.mubicdn.net/images/artworks/461678/images-original.png',
-  },
-  {
-    id: 3,
-    filmName: 'Doctor Who',
-    genre: 'Sci-fi',
-    filmReleaseDate: 1963,
-    filmImg: 'https://images.bauerhosting.com/empire/2023/10/doctor-who-main.jpg?ar=16%3A9&fit=crop&crop=top&auto=format&w=1440&q=80',
-  },
-  {
-    id: 4,
-    filmName: 'Pride',
-    genre: 'Comedy',
-    filmReleaseDate: 2014,
-    filmImg: 'https://m.media-amazon.com/images/M/MV5BMzUzMzgzMDc0Nl5BMl5BanBnXkFtZTgwMTIzMzA4MTE@._V1_.jpg',
-  },
-  {
-    id: 5,
-    filmName: 'Say Anything...',
-    genre: 'Romance',
-    filmReleaseDate: 1989,
-    filmImg: 'https://images2.minutemediacdn.com/image/upload/c_fill,w_1440,ar_16:9,f_auto,q_auto,g_auto/shape/cover/sport/sayanything-hed-728edd393579796c9826b940c944625d.jpg',
-  },
-  {
-    id: 6,
-    filmName: 'Dolittle',
-    genre: 'Comedy',
-    filmReleaseDate: 2020,
-    filmImg: 'https://cdn.mos.cms.futurecdn.net/4UtEa2Kg4iCS3uTZueeCC3-1200-80.jpg'
-  },
-  {
-    id: 7,
-    filmName: 'Dolittle',
-    genre: 'Comedy',
-    filmReleaseDate: 2020,
-    filmImg: 'https://cdn.mos.cms.futurecdn.net/4UtEa2Kg4iCS3uTZueeCC3-1200-80.jpg'
-  },
-  {
-    id: 8,
-    filmName: 'Dolittle',
-    genre: 'Comedy',
-    filmReleaseDate: 2020,
-    filmImg: 'https://cdn.mos.cms.futurecdn.net/4UtEa2Kg4iCS3uTZueeCC3-1200-80.jpg'
-  },
-  {
-    id: 9,
-    filmName: 'Say Anything...',
-    genre: 'Romance',
-    filmReleaseDate: 1989,
-    filmImg: 'https://images2.minutemediacdn.com/image/upload/c_fill,w_1440,ar_16:9,f_auto,q_auto,g_auto/shape/cover/sport/sayanything-hed-728edd393579796c9826b940c944625d.jpg',
-  },
-  {
-    id: 11,
-    filmName: 'Dolittle',
-    genre: 'Comedy',
-    filmReleaseDate: 2020,
-    filmImg: 'https://cdn.mos.cms.futurecdn.net/4UtEa2Kg4iCS3uTZueeCC3-1200-80.jpg'
-  },
-  {
-    id: 12,
-    filmName: 'Dolittle',
-    genre: 'Adventure',
-    filmReleaseDate: 2020,
-    filmImg: 'https://cdn.mos.cms.futurecdn.net/4UtEa2Kg4iCS3uTZueeCC3-1200-80.jpg'
-  },
-  {
-    id: 13,
-    filmName: 'Dolittle',
-    genre: 'Comedy',
-    filmReleaseDate: 2020,
-    filmImg: 'https://cdn.mos.cms.futurecdn.net/4UtEa2Kg4iCS3uTZueeCC3-1200-80.jpg'
-  },
-];
