--- src/mocs/full-info-film.ts
@@ -1,97 +1,0 @@
-export const fullInfoFilm = {
-  id: 1,
-  filmName: 'The Grand Budapest Hotel',
-  genre: 'Drama',
-  filmReleaseDate: 2014,
-  filmImg: 'img/bg-the-grand-budapest-hotel.jpg',
-  filmPoster: 'img/the-grand-budapest-hotel-poster.jpg',
-  runTime: '1h 39m',
-  ratingScore: 8.9,
-  ratingLevel: 'Very good',
-  ratingCount: 240,
-  description: `In the 1930s, the Grand Budapest Hotel is a popular European ski resort, presided over by concierge Gustave H. (Ralph Fiennes). Zero, a junior lobby boy, becomes Gustave's friend and protege.\n
-  Gustave prides himself on providing first-class service to the hotel's guests, including satisfying the sexual needs of the many elderly women who stay there. When one of Gustave's lovers dies mysteriously, Gustave finds himself the recipient of a priceless painting and the chief suspect in her murder.`,
-  director: 'Wes Anderson',
-  starring: ['Bill Murray', 'Edward Norton', 'Jude Law', 'Willem Dafoe',
-    'Saoirse Ronan', 'Tony Revoloru','Tilda Swinton','Tom Wilkinson','Owen Wilkinson',
-    'Adrien Brody', 'Ralph Fiennes', 'Jeff Goldblum'],
-  reviews: [
-    {
-      id: 1,
-      text: `Discerning travellers and Wes Anderson fans will luxuriate in the glorious Mittel-European 
-        kitsch of one of the director's funniest and most exquisitely designed films in years.`,
-      ratingScore: 8.9,
-      author: 'Kate Muir',
-      date: '2023-05-25T12:00:00.000Z',
-    },
-    {
-      id: 2,
-      text: `Anderson's films are too precious for some, but for those of us willing to lose ourselves in them,
-        they're a delight. "The Grand Budapest Hotel" is no different, except that he has added a hint of gravitas to the mix,
-        improving the recipe.`,
-      ratingScore: 8.0,
-      author: 'Bill Goodykoontz',
-      date: 'November 18, 2015',
-    },
-    {
-      id: 3,
-      text: `I didn't find it amusing, and while I can appreciate the creativity, 
-        it's an hour and 40 minutes I wish I could take back.`,
-      ratingScore: 8.0,
-      author: 'Amanda Greever',
-      date: 'November 18, 2015',
-    },
-    {
-      id: 4,
-      text: 'The mannered, madcap proceedings are often delightful, occasionally silly, and here and there, gruesome and/or heartbreaking.',
-      ratingScore: 7.2,
-      author: 'Matthew Lickona',
-      date: 'December 20, 2016',
-    },
-    {
-      id: 5,
-      text: 'It is certainly a magical and childlike way of storytelling, even if the content is a little more adult.',
-      ratingScore: 7.6,
-      author: 'Paula Fleri-Soler',
-      date: 'December 20, 2016',
-    },
-    {
-      id: 6,
-      text: 'It is certainly a magical and childlike way of storytelling, even if the content is a little more adult.',
-      ratingScore: 7.0,
-      author: 'Paula Fleri-Soler',
-      date: 'December 20, 2016',
-    },
-  ],
-
-  moreLikeThis:[
-    {
-      id: 10,
-      filmName: 'Fantastic Beasts: The Crimes of Grindelwald',
-      genre: 'Drama',
-      filmReleaseDate: 2017,
-      filmImg: 'img/fantastic-beasts-the-crimes-of-grindelwald.jpg',
-    },
-    {
-      id: 11,
-      filmName: 'Bohemian Rhapsody',
-      genre: 'Drama',
-      filmReleaseDate: 2017,
-      filmImg: 'img/bohemian-rhapsody.jpg',
-    },
-    {
-      id: 12,
-      filmName: 'Macbeth',
-      genre: 'Drama',
-      filmReleaseDate: 2017,
-      filmImg: 'img/macbeth.jpg',
-    },
-    {
-      id: 13,
-      filmName: 'Aviator',
-      genre: 'Drama',
-      filmReleaseDate: 2017,
-      filmImg: 'img/aviator.jpg',
-    },
-  ],
-};
