--- src/mocs/my-list.ts
@@ -1,65 +1,0 @@
-export const MyFilmList = [
-  {
-    id: 10,
-    filmName: 'Fantastic Beasts: The Crimes of Grindelwald',
-    genre: 'Drama',
-    filmReleaseDate: 2017,
-    filmImg: 'img/fantastic-beasts-the-crimes-of-grindelwald.jpg',
-  },
-  {
-    id: 11,
-    filmName: 'Bohemian Rhapsody',
-    genre: 'Drama',
-    filmReleaseDate: 2017,
-    filmImg: 'img/bohemian-rhapsody.jpg',
-  },
-  {
-    id: 12,
-    filmName: 'Macbeth',
-    genre: 'Drama',
-    filmReleaseDate: 2017,
-    filmImg: 'img/macbeth.jpg',
-  },
-  {
-    id: 13,
-    filmName: 'Aviator',
-    genre: 'Drama',
-    filmReleaseDate: 2017,
-    filmImg: 'img/aviator.jpg',
-  },
-  {
-    id: 14,
-    filmName: 'We need to talk about Kevin',
-    genre: 'Drama',
-    filmReleaseDate: 2017,
-    filmImg: 'img/we-need-to-talk-about-kevin.jpg',
-  },
-  {
-    id: 15,
-    filmName: 'What We Do in the Shadows',
-    genre: 'Drama',
-    filmReleaseDate: 2017,
-    filmImg: 'img/what-we-do-in-the-shadows.jpg',
-  },
-  {
-    id: 16,
-    filmName: 'Revenant',
-    genre: 'Drama',
-    filmReleaseDate: 2017,
-    filmImg: 'img/revenant.jpg',
-  },
-  {
-    id: 17,
-    filmName: 'Johnny English',
-    genre: 'Drama',
-    filmReleaseDate: 2017,
-    filmImg: 'img/johnny-english.jpg',
-  },
-  {
-    id: 18,
-    filmName: 'Shutter Island',
-    genre: 'Drama',
-    filmReleaseDate: 2017,
-    filmImg: 'img/shutter-island.jpg',
-  },
-];
