+++ src/store/index.ts
@@ -1,21 +1,22 @@
-import { configureStore } from '@reduxjs/toolkit';
+import { combineReducers, configureStore } from '@reduxjs/toolkit';
 import { createAPI } from '../api/api';
-import { userReducer } from './reducers/user-reducer';
-import { fullFilmReducer } from './reducers/film-reducer';
-import { reviewReducer } from './reducers/review-reducer';
-import { filmsReducer } from './reducers/films-reducer';
-import { favoritesReducer } from './reducers/favorite-reducer';
+import { NameSpace } from '../const';
+import { filmsSlice } from './slices/films-slice';
+import { fullFilmSlice } from './slices/film-slice';
+import { favoritesSlice } from './slices/favorites-slice';
+import { userSlice } from './slices/user-slice';
+import { reviewSlice } from './slices/review-slice';
 
 
 const api = createAPI();
 
-const reducer = {
-  films: filmsReducer,
-  fullFilm: fullFilmReducer,
-  favorites:favoritesReducer,
-  user: userReducer,
-  reviews: reviewReducer,
-};
+const reducer = combineReducers({
+  [NameSpace.Films]: filmsSlice.reducer,
+  [NameSpace.Film]: fullFilmSlice.reducer,
+  [NameSpace.MyFilms]: favoritesSlice.reducer,
+  [NameSpace.User]: userSlice.reducer,
+  [NameSpace.Review]: reviewSlice.reducer,
+});
 
 export const store = configureStore({
   reducer,
